[Feat] [FE-54] 로딩 상태 확인 후 링킹
