+++ src/navigations/index.tsx
@@ -1,11 +1,17 @@
-import { useEffect } from 'react';
+import { View, Text } from 'react-native';
+import { useEffect, useState } from 'react';
 import 'react-native-gesture-handler';
+import AsyncStorage from '@react-native-async-storage/async-storage';
+import { useRecoilState } from 'recoil';
 import {
   NativeStackNavigationProp,
   createNativeStackNavigator,
 } from '@react-navigation/native-stack';
-
-import { NavigatorScreenParams } from '@react-navigation/native';
+import * as Linking from 'expo-linking';
+import {
+  NavigationContainer,
+  NavigatorScreenParams,
+} from '@react-navigation/native';
 import MainBottomNavTab, {
   MainBottomNavTabParamList,
 } from './MainBottomNavTab/MainBottomNavTab';
@@ -13,7 +19,6 @@ import CategoryScreen from '@/screens/Category/CategoryScreen';
 import Colors from '@/styles/colors';
 import SearchNotificationRight from '@/components/molecules/Header/SearchNotificationRight';
 import BackLeft from '@/components/molecules/Header/BackLeft';
-import useLoginHook from '@/hooks/loginHook';
 
 import AuthStack, { AuthStackParamList } from './AuthStack/AuthStack';
 import TermsStack, { TermsStackParamList } from './Common/TermsStack';
@@ -24,6 +29,8 @@ import RightCloseX from '@/components/molecules/Header/RightCloseX';
 import { Category, CategoryFindLabel } from '@/constants/category';
 import ProductListScreen from '@/screens/Product/ProductListScreen';
 import Fonts from '@/styles/typography';
+import ModalProvider from '@/components/organisms/Modal/ModalProvider';
+import loginState from '@/utils/recoil/login';
 
 export type RootStackParamList = {
   MainBottomNavTab: NavigatorScreenParams<MainBottomNavTabParamList>;
@@ -40,88 +47,130 @@ export type RootStackNavigationProp =
 
 const Stack = createNativeStackNavigator<RootStackParamList>();
 
+const prefix = Linking.createURL('/');
+
 function Navigation() {
-  const { isLoggedIn, getStoredToken } = useLoginHook();
+  // const { isLoggedIn, getStoredToken } = useLoginHook();
+  const [loading, setLoading] = useState(true);
+  const [isLoggedIn, setIsLoggedIn] = useRecoilState(loginState);
+
+  const getStoredToken = async () => {
+    const token = await AsyncStorage.getItem('token');
+    if (token) {
+      setIsLoggedIn(true);
+    }
+    setLoading(false);
+  };
+
+  const linking = {
+    prefixes: [prefix],
+    config: {
+      screens: {
+        ContentStack: {
+          path: '',
+          screens: {
+            Content: {
+              path: 'content/:id',
+              parse: {
+                id: Number,
+              },
+            },
+          },
+        },
+      },
+    },
+  };
 
   useEffect(() => {
     getStoredToken();
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, []);
 
-  return (
-    <Stack.Navigator>
-      {isLoggedIn ? (
-        <>
-          {/* bottom nav */}
-          <Stack.Screen
-            name='MainBottomNavTab'
-            component={MainBottomNavTab}
-            options={{ headerShown: false }}
-          />
-          {/* category */}
-          <Stack.Screen
-            name='Category'
-            component={CategoryScreen}
-            options={({ route }) => ({
-              title: CategoryFindLabel(route.params.categoryId),
-              headerStyle: {
-                backgroundColor: Colors.primary,
-              },
-              headerTintColor: '#fff',
-              headerTitleStyle: {
-                fontFamily: 'Montserrat-ExtraBold',
-                fontSize: 20,
-              },
-              headerLeft: () => <BackLeft theme='white' />,
-              headerRight: () => <SearchNotificationRight theme='white' />,
-            })}
-          />
-          {/* product list */}
-          <Stack.Screen
-            name='ProductList'
-            component={ProductListScreen}
-            options={({ route }) => ({
-              title: route.params.title,
-              headerLeft: () => <BackLeft />,
-              headerTitleStyle: Fonts.header.title,
-            })}
-          />
-          {/* content */}
-          <Stack.Screen
-            name='ContentStack'
-            component={ContentStack}
-            options={{ headerShown: false }}
-          />
+  if (loading && !isLoggedIn) {
+    return (
+      <View style={{ flex: 1, justifyContent: 'center' }}>
+        <Text>기다리는중</Text>
+      </View>
+    );
+  }
 
-          {/* Notification */}
+  return (
+    <NavigationContainer linking={linking}>
+      <Stack.Navigator>
+        {isLoggedIn ? (
+          <>
+            {/* bottom nav */}
+            <Stack.Screen
+              name='MainBottomNavTab'
+              component={MainBottomNavTab}
+              options={{ headerShown: false }}
+            />
+            {/* category */}
+            <Stack.Screen
+              name='Category'
+              component={CategoryScreen}
+              options={({ route }) => ({
+                title: CategoryFindLabel(route.params.categoryId),
+                headerStyle: {
+                  backgroundColor: Colors.primary,
+                },
+                headerTintColor: '#fff',
+                headerTitleStyle: {
+                  fontFamily: 'Montserrat-ExtraBold',
+                  fontSize: 20,
+                },
+                headerLeft: () => <BackLeft theme='white' />,
+                headerRight: () => <SearchNotificationRight theme='white' />,
+              })}
+            />
+            {/* product list */}
+            <Stack.Screen
+              name='ProductList'
+              component={ProductListScreen}
+              options={({ route }) => ({
+                title: route.params.title,
+                headerLeft: () => <BackLeft />,
+                headerTitleStyle: Fonts.header.title,
+              })}
+            />
+            {/* content */}
+            <Stack.Screen
+              name='ContentStack'
+              component={ContentStack}
+              options={{ headerShown: false }}
+            />
+
+            {/* Notification */}
+            <Stack.Screen
+              name='Notification'
+              component={NotificationScreen}
+              options={{
+                headerTitle: '',
+                headerLeft: () => <HeaderLeftTitle title='Notification' />,
+                headerRight: () => <RightCloseX />,
+                headerShadowVisible: false,
+              }}
+            />
+          </>
+        ) : (
           <Stack.Screen
-            name='Notification'
-            component={NotificationScreen}
+            name='AuthStack'
+            component={AuthStack}
             options={{
-              headerTitle: '',
-              headerLeft: () => <HeaderLeftTitle title='Notification' />,
-              headerRight: () => <RightCloseX />,
-              headerShadowVisible: false,
+              headerShown: false,
             }}
           />
-        </>
-      ) : (
-        <Stack.Screen
-          name='AuthStack'
-          component={AuthStack}
-          options={{
-            headerShown: false,
-          }}
-        />
-      )}
+        )}
 
-      {/* terms */}
-      <Stack.Screen
-        name='TermsStack'
-        component={TermsStack}
-        options={{ headerShown: false }}
-      />
-    </Stack.Navigator>
+        {/* terms */}
+        <Stack.Screen
+          name='TermsStack'
+          component={TermsStack}
+          options={{ headerShown: false }}
+        />
+      </Stack.Navigator>
+      <ModalProvider />
+    </NavigationContainer>
   );
 }
